requirements.txt
# Web frameworks
Django==5.1.15
djangorestframework==3.15.2
fastapi==0.127.0
uvicorn==0.40.0

# WSGI/production server
gunicorn==23.0.0

# Authentication / security
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
cryptography==45.0.0

# Database / ORM utilities
SQLAlchemy==2.0.36

# Caching / async tasks
redis==5.2.1
celery==5.4.0

# HTTP / networking
requests==2.32.3
urllib3==2.3.0

# Environment management
python-dotenv==1.2.2

# CORS / API utilities
django-cors-headers==4.6.0

# Schema / API docs
drf-spectacular==0.29.0
drf-yasg==1.21.8

# Google APIs (keep only if used)
google-api-python-client==2.161.0
google-auth==2.38.0
google-cloud-storage==3.7.0
google-generativeai==0.8.4

# AWS SDK (keep only if used)
boto3==1.42.16
botocore==1.42.16

# Utilities
pydantic==2.10.6
pydantic-settings==2.6.1
pandas==2.2.2
numpy==1.26.4

# File handling
openpyxl==3.1.5

# Logging / monitoring (optional but common)
sentry-sdk==2.19.2

whitenoise==6.12.0
django-redis==5.4.0
bcrypt==4.1.2
itsdangerous==2.2.0